requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
celery
redis
PyMuPDF
pytesseract
pdf2image
openai
pinecone
python-dotenv
whitenoise
langchain
langchain-openai
langchain-pinecone